requirements.txt
discord.py
asyncio
socket